IDL-88 WIP fix cloud services tests
