+++ openidl-common-lib/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@senofi/openidl-common-lib",
-  "version": "0.7.70",
+  "version": "0.7.71",
   "private": false,
   "main": "index.js",
   "publishConfig": {
