+++ openidl-common-lib/test/cloud-services/secret/secrets-client-factory.test.js
@@ -1,19 +1,19 @@
 const {expect} = require('chai');
-const AWSSecretsManagerClient = require('../../secret/impl/aws-secrets-manager-client');
-const AzureKeyVaultClient = require('../../secret/impl/azure-key-vault-secrets-client');
-const KubernetesClient = require('../../secret/impl/kubernetes-secrets-client');
-const secretsStoreType = require('../../constants/secrets-store-type');
+const AWSSecretsManagerClient = require('../../../cloud-services/secret/impl/aws-secrets-manager-client');
+const AzureKeyVaultClient = require('../../../cloud-services/secret/impl/azure-key-vault-secrets-client');
+const KubernetesClient = require('../../../cloud-services/secret/impl/kubernetes-secrets-client');
+const secretsStoreType = require('../../../cloud-services/constants/secrets-store-type');
 
 describe('SecretsClientFactory resolving secrets client AWS environment', () => {
     before(() => {
         process.env['KVS_CONFIG'] = JSON.stringify({ secretsStoreType: secretsStoreType.AWS_SECRETS_MANAGER });
     })
     after(() => {
-        delete require.cache[require.resolve('../../secret/secrets-client-factory')];
+        delete require.cache[require.resolve('../../../cloud-services/secret/secrets-client-factory')];
         process.env['KVS_CONFIG'] = JSON.stringify({})
     });
     it('should return the same AWSSecretsManagerClient instance for the same environment twice', async () => {
-        const SecretsClientFactory = require('../../secret/secrets-client-factory');
+        const SecretsClientFactory = require('../../../cloud-services/secret/secrets-client-factory');
         const secretsClient1 = SecretsClientFactory.getInstance();
         const secretsClient2 = SecretsClientFactory.getInstance();
 
@@ -33,12 +33,12 @@ describe('SecretsClientFactory resolving secrets client Azure environment', () =
             azureVaultUrl: 'mockedVaultUrl'});
     })
     after(() => {
-        delete require.cache[require.resolve('../../secret/secrets-client-factory')];
+        delete require.cache[require.resolve('../../../cloud-services/secret/secrets-client-factory')];
         process.env['KVS_CONFIG'] = JSON.stringify({})
     });
 
     it('should return the same AzureKeyVaultClient instance for the same environment twice', async () => {
-        const SecretsClientFactory = require('../../secret/secrets-client-factory');
+        const SecretsClientFactory = require('../../../cloud-services/secret/secrets-client-factory');
         const secretsClient1 = SecretsClientFactory.getInstance();
         const secretsClient2 = SecretsClientFactory.getInstance();
 
@@ -54,12 +54,12 @@ describe('SecretsClientFactory resolving secrets client Kubernetes environment',
     });
 
     after(() => {
-        delete require.cache[require.resolve('../../secret/secrets-client-factory')];
+        delete require.cache[require.resolve('../../../cloud-services/secret/secrets-client-factory')];
         process.env['KVS_CONFIG'] = JSON.stringify({})
     });
 
     it('should return the same KubernetesClient instance for the same environment twice', async () => {
-        const SecretsClientFactory = require('../../secret/secrets-client-factory');
+        const SecretsClientFactory = require('../../../cloud-services/secret/secrets-client-factory');
         const secretsClient1 = SecretsClientFactory.getInstance();
         const secretsClient2 = SecretsClientFactory.getInstance();
